+++ CHANGELOG.md
@@ -13,6 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities
 
+## [3.2.0] - Upcoming
+
+### Added
+
+- Added support for `data-skin-current-accordion` attribute
+- Added support for relative URL's when using `data-skin-current-class`
+
 ## [3.1.3] - 2025.11.05
 
 ### Added